Conflicts:
	app/models/marty/api_auth.rb
	app/models/marty/import_type.rb
	app/models/marty/posting.rb
	app/models/marty/posting_type.rb
	app/models/marty/role.rb
	app/models/marty/script.rb
	app/models/marty/tag.rb
	app/models/marty/user_role.rb
	marty.gemspec

+++ app/models/marty/token.rb
@@ -22,7 +22,7 @@ private
 
   # Removes obsolete tokens
   def delete_previous_tokens
-    if user
+    if user && !Marty::Util.db_in_recovery?
       Token.delete_all(['user_id = ?', user.id])
     end
   end
